# backend/docs/swagger.yaml
definitions:
...
  handler.APIResponse:
    properties:
      data: {}
      msg:
        example: ok
        type: string
    type: object
...
  handler.bLoginRequest:
    properties:
      password:
        type: string
      phone:
        type: string
    required:
    - password
    - phone
    type: object
  handler.bRefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.bannerRequest:
    properties:
      image_url:
        type: string
      link_type:
        type: string
      link_value:
        type: string
      sort:
        type: integer
      station_id:
        type: integer
      status:
        type: string
      title:
        type: string
    required:
    - image_url
    type: object
...
info:
  contact: {}
paths:
...
  /b/auth/login:
    post:
      consumes:
      - application/json
      description: B端工作人员使用手机号和密码登录
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.bLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回token和用户信息
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 用户已禁用或无B端角色
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: B端用户登录
      tags:
      - b_auth
...
  /b/auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录的B端用户详细信息及权限列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户信息和权限
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: 无B端角色
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取当前B端用户信息
      tags:
      - b_auth
  /b/auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.bRefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: 令牌无效或已过期
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 刷新访问令牌
      tags:
      - b_auth
  /b/banners:
    get:
      consumes:
      - application/json
      description: 分页获取轮播图列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 站点ID筛选
        in: query
        name: station_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 获取轮播图列表（管理端）
      tags:
      - b_banner
    post:
      consumes:
      - application/json
      description: 创建新的轮播图
      parameters:
      - description: 轮播图信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.bannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 创建轮播图
      tags:
      - b_banner
  /b/banners/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定轮播图
      parameters:
      - description: 轮播图ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 删除轮播图
      tags:
      - b_banner
    put:
      consumes:
      - application/json
      description: 更新指定轮播图信息
      parameters:
      - description: 轮播图ID
        in: path
        name: id
        required: true
        type: integer
      - description: 轮播图信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.bannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - Bearer: []
      summary: 更新轮播图
      tags:
      - b_banner
...
  /c/banners:
    get:
      consumes:
      - application/json
      description: 获取当前站点或全局的轮播图
      parameters:
      - description: 站点ID（0或不传表示全局）
        in: query
        name: station_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 获取轮播图列表
      tags:
      - public
...
swagger: "2.0"